project/requirements.txt
djangorestframework
django
django-webpack-loader